So many updates!
If you want to run the current version of the website on your own machine. Here's what you need to do:
navigate to /frontend and then run: npm install

This will install all of the necessary dependencies on your machine! Going forward try to make your components work with this version of the website before making a pull request.